Separate execution loop to standalone component
