+++ src-tests/Console/Command/RunCommandIntegrationTest.php
@@ -12,6 +12,7 @@ use Symfony\Component\EventDispatcher\EventDispatcher;
  * Run command tests that require real Selenium server to execute the tests.
  *
  * @covers \Lmc\Steward\Console\Command\RunCommand
+ * @covers \Lmc\Steward\Process\ExecutionLoop
  * @group integration
  * @runTestsInSeparateProcesses
  */
